requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
altair=5.0.1=pypi_0
attrs=23.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
basicsr=1.3.4.9=pypi_0
beautifulsoup4=4.12.3=pypi_0
blinker=1.6.3=pypi_0
ca-certificates=2023.08.22=haa95532_0
cachetools=5.3.2=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
cycler=0.11.0=pypi_0
cython=3.0.9=pypi_0
decorator=5.1.1=pypi_0
dpcpp-cpp-rt=2024.0.2=pypi_0
einops=0.6.1=pypi_0
filelock=3.12.2=pypi_0
fonttools=4.38.0=pypi_0
fsspec=2023.1.0=pypi_0
future=0.18.3=pypi_0
gdown=4.7.3=pypi_0
git=2.40.1=haa95532_1
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
google-auth=2.27.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.60.0=pypi_0
h5py=3.8.0=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.6=pypi_0
imageio=2.31.2=pypi_0
importlib-metadata=6.7.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-cmplr-lib-rt=2024.0.2=pypi_0
intel-cmplr-lic-rt=2024.0.2=pypi_0
intel-opencl-rt=2024.0.2=pypi_0
intel-openmp=2024.0.2=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.5=pypi_0
lmdb=1.4.1=pypi_0
lpips=0.1.4=pypi_0
markdown=3.4.4=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.4=pypi_0
matplotlib=3.5.3=pypi_0
mdurl=0.1.2=pypi_0
mkl=2024.0.0=pypi_0
mkl-fft=1.3.1=pypi_0
mkl-random=1.2.2=pypi_0
mkl-service=2.4.0=pypi_0
natsort=8.4.0=pypi_0
networkx=2.6.3=pypi_0
numpy=1.21.6=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=1.1.1w=h2bbff1b_0
packaging=23.2=pypi_0
pandas=1.3.5=pypi_0
pillow=9.3.0=pypi_0
pip=22.3.1=py37haa95532_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.0.0=pypi_0
protobuf=3.20.3=pypi_0
pyarrow=12.0.1=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.1.1=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=pypi_0
python=3.7.16=h6244533_0
python-dateutil=2.8.2=pypi_0
pytz=2024.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0.1=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.7.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setuptools=65.6.3=py37haa95532_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h2bbff1b_0
streamlit=1.23.1=pypi_0
streamlit-lottie=0.0.3=pypi_0
streamlit-option-menu=0.3.2=pypi_0
tb-nightly=2.12.0a20230113=pypi_0
tbb=2021.11.0=pypi_0
tenacity=8.2.3=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.9.12=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.1=pypi_0
torch=1.8.1+cu111=pypi_0
torchaudio=0.8.1=pypi_0
torchvision=0.9.1+cu111=pypi_0
tornado=6.2=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2024.1=pypi_0
tzlocal=4.3.1=pypi_0
urllib3=2.0.7=pypi_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37haa95532_0
wincertstore=0.2=py37haa95532_2
yacs=0.1.8=pypi_0
yapf=0.40.2=pypi_0
zipp=3.15.0=pypi_0
